Revert "[clang][ARM][AArch64] Define intrinsics guarded by __has_builtin on a…"
This reverts commit de0153da325b21340747d8a064906a8c1c2f95ea.
